App now handles signing in to X accounts that need verification
When the app tries to sign in to an X account that needs verification,
a message is displayed stating that the X account should have no
verification. The user can then edit the email, username and password
and start the bot again.

Also, when trying to sign in, this app handles when X asks for a phone
number or email. The app provides an email when this happens.